refactor(CQ2-2): split bangumi_api.py (1044 lines) into bangumi_api/ package
Split the monolithic BangumiApi client into a 5-file mixin-based package:
- __init__.py: BangumiApi facade (init/get/post/put/patch + cache)
- http_layer.py: HttpLayerMixin (retry, proxy fallback, network diagnosis)
- search.py: SearchMixin (search/subject/related/bgm_search/title_diff)
- episodes.py: EpisodesMixin (episode resolution by sort/airdate/sequel)
- collection.py: CollectionMixin (subject/episode collection state)

Re-export httpx/socket/time from __init__.py for test patch compatibility
(app.utils.bangumi_api.httpx.Client / socket.getaddrinfo / time.sleep).

All 1894 tests pass, ruff clean.
